Refactor game: separate scenes, organize assets
