+++ src/game/assets/bg.png
